% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_datasets_script_bug_bug.R
\name{func_generate_spike_structure}
\alias{func_generate_spike_structure}
\title{Generate the spike structure as a list, and the correlation matrix}
\usage{
func_generate_spike_structure(range_features, domain_features, num_features,
  log_corr_values)
}
\arguments{
\item{range_features}{A vector of features which will be replaced with
spiked data}

\item{domain_features}{A vector which contains the features that will be
used to generate the spiked data}

\item{num_features}{is the number of features in the dataset
(so the dimension of the correlation matrix)}

\item{log_corr_values}{is a scalar or vector of correlation values. If
one value, all spikes will have that correlation value. If a vector
of length > \code{num_features}, it will be sampled without
replacement. If a vector of length < \code{num_features}, it will be
recycled.}
}
\value{
A correlation matrix corresponding to the spike structure.
}
\description{
Generate the spike structure as a list, and the correlation matrix
}

