% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_datasets_script_bug_bug.R
\name{func_get_log_corr_mat_from_file}
\alias{func_get_log_corr_mat_from_file}
\title{Get the log-basis correlation matrix from a file}
\usage{
func_get_log_corr_mat_from_file(num_features, file_name)
}
\arguments{
\item{num_features}{is the number of features in the dataset
(so the dimension of the correlation matrix)}

\item{file_name}{The name of the file where the correlation values are
stored. Should have fields `Domain`, `Range`, and `Correlation`.}
}
\value{
A list contains the microbe-microbe spiked abundance matrix and the 
    information about the pairs of microbes that are correlated when the pairs
    are specified through a user provided file.
}
\description{
Get the log-basis correlation matrix from a file
}

