% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_datasets_script_bug_bug.R
\name{func_get_log_corr_mat_from_num}
\alias{func_get_log_corr_mat_from_num}
\title{A function to generate a bug-bug correlation matrix with a
    fixed number of spikes.}
\usage{
func_get_log_corr_mat_from_num(num_features, num_spikes, log_corr_values)
}
\arguments{
\item{num_features}{is the number of features in the dataset
(so the dimension of the correlation matrix)}

\item{num_spikes}{is the number of non-zero correlations to
generate}

\item{log_corr_values}{is a scalar or vector of correlation values. If
one value, all spikes will have that correlation value. If a vector
of length > \code{num_features}, it will be sampled without
replacement. If a vector of length < \code{num_features}, it will be
recycled.}
}
\value{
A list contains the microbe-microbe spiked abundance matrix and the 
    information about the pairs of microbes that are correlated when only 
    number of correlated microbes are specified.
}
\description{
A function to generate a bug-bug correlation matrix with a
    fixed number of spikes.
}

