% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_datasets_script_bug_bug.R
\name{func_list_to_str}
\alias{func_list_to_str}
\title{Convert a list of vectors into an appropriately-delimited single string}
\usage{
func_list_to_str(mylist)
}
\arguments{
\item{mylist}{A list of vectors}
}
\value{
A string concatenates all elements in a list by semi-colons.
}
\description{
Convert a list of vectors into an appropriately-delimited single string
}

